Added sprites, seperated scripts, added setup func
Created some super simple sprites on piskel and added that sprite sheet
to the game, also updated the script to use the new sprites from the
sheet. I also seperated out the scripts into their various components
since everything was previously inside of createGrid, there is now a
vision, grid, and PlayerController class. The grid class simply contains
the 2D List of Tiles that makes up the map of the game, it also serves as
the access point for map based functions like GetNorthTile(tile).
PlayerController controls the players movement and handles creating a
object of the Vision class which contains the visibility methods. It also handles
calling the visibility methods as the player moves. In addition
PlayerController now contains a SetUp method which rotates the player sprite
the correct direction and reveals all of the correct walls during
startup.