Revamped player vision, graphics, and camera
The player vision is now circular based, in other words the player can
see through walls but only to a certain range. This is to allow the game
play to work when the player can push the walls of the maze later on. I
have also left the old line based vision code because this should be
useful for the enemy AI code. The graphics were revamped mainly to allow
the circular camera setup to work. All tiles now have a originalSprite
and originalColor in order to reset them once the player moves past.
Invisible tiles are now set to the camera's back ground color of black
making them blend in and giving a strong circle look to our players FOV.
Finally the playerMovement code was updated in order to allow the camera
to follow the player and also to allow in place rotations in preperation
for adding wall pushing mechanics.