Start of wall pushing ability
Player can now push walls into other tiles as long as the wall has a
floor tile to move onto. The player has to press twice into a wall to
move it, which serves two goals later this means it will cost a turn to
"brace" yourself against the wall, and it also prevents against
accidental wall pushes.